feat(configuration): 现在默认配置会与从配置文件中读取的配置信息合并. 支持从配置文件中自定义管理员账号密码.

+++ internal/pkg/configuration/read_arguments.go
@@ -16,7 +16,7 @@ func parseArguments() *Configuration {
 
 	return &Configuration{
 		ServerMonitor: ServerMonitorConfiguration{
-			Port: uint(*serverPort),
+			Port: *serverPort,
 			Mock: *mockMode,
 		}}
 }
